requirements.txt
Babel==2.14.0
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
future==1.0.0
iso8601==2.1.0
kiwisolver==1.4.5
matplotlib==3.8.4
mysql-connector-python==8.3.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.1
serial==0.0.97
six==1.16.0
tk==0.1.0
tkcalendar==1.6.1
